# www/public/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "description": "Issued when a browser pairs with the daemon.",
        "in": "cookie",
        "name": "picode_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP API the PiCode web UI itself speaks — same daemon, same pairing-gated session. Generated from the server's route registration by cmd/picode-openapi; do not edit the JSON by hand. Authenticate with the picode_session cookie issued when you pair a browser (Settings → Pairing).",
    "title": "PiCode HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents": {
      "get": {
        "operationId": "getApiAgents",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "postApiAgents",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{id}": {
      "delete": {
        "operationId": "deleteApiAgents",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "patchApiAgents",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/agents/{id}/cleanup": {
      "get": {
        "operationId": "getApiAgentsCleanup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/agents/{id}/close": {
      "post": {
        "operationId": "postApiAgentsClose",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/agents/{id}/open": {
      "post": {
        "operationId": "postApiAgentsOpen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/agents/{id}/tasks": {
      "get": {
        "operationId": "getApiAgentsTasks",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "postApiAgentsTasks",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/server": {
      "get": {
        "operationId": "getApiServer",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "server"
        ]
      }
    },
...
    "/api/workspaces": {
      "get": {
        "operationId": "getApiWorkspaces",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "postApiWorkspaces",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/workspaces/{id}": {
      "delete": {
        "operationId": "deleteApiWorkspaces",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/agents": {
      "post": {
        "operationId": "postApiWorkspacesAgents",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/workspaces/{id}/cleanup": {
      "get": {
        "operationId": "getApiWorkspacesCleanup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/close": {
      "post": {
        "operationId": "postApiWorkspacesClose",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/favicon": {
      "get": {
        "operationId": "getApiWorkspacesFavicon",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/workspaces/{id}/open": {
      "post": {
        "operationId": "postApiWorkspacesOpen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/workspaces/{id}/sessions": {
      "get": {
        "operationId": "getApiWorkspacesSessions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/workspaces/{id}/sessions/new": {
      "post": {
        "operationId": "postApiWorkspacesSessionsNew",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/sessions/rename": {
      "post": {
        "operationId": "postApiWorkspacesSessionsRename",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/sessions/resume": {
      "post": {
        "operationId": "postApiWorkspacesSessionsResume",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/sessions/transcript": {
      "get": {
        "operationId": "getApiWorkspacesSessionsTranscript",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{id}/status": {
      "get": {
        "operationId": "getApiWorkspacesStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "workspaces"
        ]
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "local daemon",
      "url": "http://127.0.0.1:8445"
    }
  ],
...
}